requirements.txt
setproctitle>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow
mss>=9.0.0
dxcam[cv2,winrt]
flask>=3.0.0
pyserial>=3.5
